requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==1.10.13
supabase==2.3.0
httpx==0.24.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.3
